+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 Select a Spreadsheet Light version below to view the changelog history:
 
-- [Spreadsheet Light 0.2.0](#V0.2.0) **Current**
+- [Spreadsheet Light 0.2.0](#V0.2.1) **Current**
+- [Spreadsheet Light 0.2.0](#V0.2.0)
 - [Spreadsheet Light 0.1.11](#V0.1.11)
 - [Spreadsheet Light 0.1.10](#V0.1.10)
 - [Spreadsheet Light 0.1.9](#V0.1.9)
 - [Spreadsheet Light 0.1.7](#V0.1.7)
 - [Spreadsheet Light 0.1.6](#V0.1.6)
+
+## V0.2.1
+
+Adding support for updating the content within the same object with a new CSV string. To this function can be subscribed too.
 
 ## V0.2.0